list_all_stacks/requirements.txt
keystoneauth1==2.21.0
python-keystoneclient==3.11.0
python-heatclient==1.10.0
